requirements.txt
fastapi==0.115.6
uvicorn==0.32.0
httpx==0.27.2
pydantic==2.9.2
PyJWT==2.9.0
PyYAML==6.0.2
jinja2==3.1.4
python-multipart==0.0.12
rich==13.9.0
